przepisanie tłumaczeń do jednego pliku
