+++ fullstack-app/frontend/src/assets/translations.ts
@@ -1,0 +1,118 @@
+const translations = {
+  en: {
+    login: "login",
+    loginButton: "Login",
+    password: "password",
+    passwordBig: "Password",
+    register: "Register",
+    loginSuccess: "Login successful!",
+    unknownError: "An unknown error occurred.",
+    noAccountQuestion: "Don't have an account?",
+    userProfile: "User profile",
+    editProfile: "Edit profile",
+    saveChanges: "Save changes",
+    returnToDashboard: "Return to dashboard",
+    cancel: "cancel",
+    cancelBig: "Cancel",
+    newPassword: "New password",
+    noTokenError: "No token found. Cannot fetch profile.",
+    updatePrifleSuccess: "Profile updated successfully!",
+    updateProfileError: "Failed to update profile. Please try again.",
+    email: "email",
+    username: "username",
+    repeatPassword: "repeat password",
+    registerButton: "Register",
+    accountCreated: "Account created successfully! You can now log in.",
+    passwordsDoNotMatch: "Passwords do not match!",
+    accountAlreadyExistsQuestion: "Already have an account?",
+    settings: "settings",
+    settingBigs: "Settings",
+    customize: "Customize your experience",
+    theme: "Theme",
+    language: "Language",
+    fontSize: "Font Size",
+    preview: "Preview",
+    fontSizes: ["small", "medium", "large"],
+    themes: ["light", "dark"],
+    notes: "notes",
+    profile: "user profile",
+    logout: "log out",
+    select: "Select a note to view or edit its content.",
+    notLoggedError: "You must be logged in to save notes.",
+    noNoteIDerror: "Note ID is not set. Cannot update note.",
+    noteUpdateSuccess: "Note updated successfully!",
+    noteUpdateError: "Failed to update note. Please try again.",
+    save: "save",
+    create: "create Note",
+    delete: "delete Note",
+    changeName: "change Name",
+    noteTitle: "Note Title",
+    noteTitlePlaceholder: "Enter note title",
+    noteCreated: "Note created successfully.",
+    noteDeleted: "Note deleted successfully.",
+    noteTitleUpdated: "Note title updated successfully.",
+    notLoggedIn: "You must be logged in to perform this action.",
+    emptyTitle: "Note title cannot be empty.",
+    noteCreateError: "Failed to create note. Please try again.",
+    noteDeleteError: "Failed to delete note. Please try again.",
+  },
+  pl: {
+    login: "login",
+    loginButton: "Zaloguj się",
+    password: "hasło",
+    passwordBig: "Hasło",
+    register: "Zarejestruj się",
+    loginSuccess: "Logowanie powiodło się!",
+    unknownError: "Wystąpił nieznany błąd.",
+    noAccountQuestion: "Nie masz konta?",
+    userProfile: "Profil użytkownika",
+    editProfile: "Edytuj profil",
+    saveChanges: "Zapisz zmiany",
+    returnToDashboard: "Powrót na główny panel",
+    cancel: "anuluj",
+    cancelBig: "Anuluj",
+    newPassword: "Nowe hasło",
+    noTokenError: "Brak tokenu. Nie można pobrać profilu.",
+    updatePrifleSuccess: "Profil został pomyślnie zaktualizowany!",
+    updateProfileError: "Nie udało się zaktualizować profilu. Spróbuj ponownie.",
+    email: "email",
+    username: "nazwa użytkownika",
+    repeatPassword: "powtórz hasło",
+    registerButton: "Zarejestruj się",
+    accountCreated: "Konto zostało pomyślnie utworzone! Możesz się teraz zalogować.",
+    passwordsDoNotMatch: "Hasła nie pasują do siebie!",
+    accountAlreadyExistsQuestion: "Masz już konto?",
+    settings: "ustawienia",
+    settingBigs: "Ustawienia",
+    customize: "Dostosuj swoje doświadczenie",
+    theme: "Motyw",
+    language: "Język",
+    fontSize: "Rozmiar czcionki",
+    preview: "Podgląd",
+    fontSizes: ["mały", "średni", "duży"],
+    themes: ["jasny", "ciemny"],
+    notes: "notatki",
+    profile: "profil użytkownika",
+    logout: "wyloguj",
+    select: "Wybierz notatkę aby zobaczyć i edytować jej zawartość.",
+    notLoggedError: "Mysisz być zalogowany, aby zapisać notatki.",
+    noNoteIDerror: "ID notatki nie jest ustawione. Nie można zaktualizować notatki.",
+    noteUpdateSuccess: "Notatka została pomyślnie zaktualizowana!",
+    noteUpdateError: "Nie udało się zaktualizować notatki. Spróbuj ponownie.",
+    save: "zapisz",
+    create: "utwórz notatkę",
+    delete: "usuń notatkę",
+    changeName: "zmień nazwę",
+    noteTitle: "Tytuł notatki",
+    noteTitlePlaceholder: "Wprowadź tytuł notatki",
+    noteCreated: "Notatka została pomyślnie utworzona.",
+    noteDeleted: "Notatka została pomyślnie usunięta.",
+    noteTitleUpdated: "Tytuł notatki został pomyślnie zaktualizowany.",
+    notLoggedIn: "Musisz być zalogowany, aby wykonać tę czynność.",
+    emptyTitle: "Tytuł notatki nie może być pusty.",
+    noteCreateError: "Nie udało się utworzyć notatki. Spróbuj ponownie.",
+    noteDeleteError: "Nie udało się usunąć notatki. Spróbuj ponownie.",
+  },
+}
+
+export default translations;
